requirements.txt
Django==1.11.3
gunicorn==19.9.0
django-cors-headers==2.1.0
django-crispy-forms==1.6.1
django-filter==1.1.0
django-formtools==2.1
django-guardian==1.4.9
django-import-export==0.6.1
django-pure-pagination==0.3.0
django-ranged-response==0.2.0
django-redis==4.9.0
django-reset-migrations==0.3.1
django-reversion==2.0.12
django-simple-captcha==0.5.6
djangorestframework==3.7.7
djangorestframework-jwt==1.11.0
mysqlclient==1.3.12